refactor(warp): switch to implementing ReportSink interface PE-4862
The ObservationPublisher interface was added before the ReportSink
interface. Now that we have the ReportSink interface we might as well
consolidate to it since it's a little more general.

Note: There's still some work to do to finalize how save results are
passed between ReportSink implementations.
